+++ examples/regional_private_node_pool_oauth_scopes/provider.tf
@@ -15,9 +15,9 @@
  */
 
 provider "google" {
-  version = "2.18.0"
+  version = "3.3.0"
 }
 
 provider "google-beta" {
-  version = "2.18.0"
+  version = "3.3.0"
 }