+++ modules/safer-cluster/versions.tf
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+// This file was automatically generated from a template in ./autogen/safer-cluster
+
 terraform {
   required_version = ">= 0.12"
 }